refactor(github-issues-runtime-tests): extract chat control module (#1272)
